Trait rsp2_lammps_wrap::Potential

pub trait Potential {
    type Meta: Clone;
    fn init_info(&self, coords: &Coords, meta: &Self::Meta) -> InitInfo;

    fn atom_types(&self, coords: &Coords, meta: &Self::Meta) -> Vec<AtomType>;

    fn molecule_ids(
        &self, 
        coords: &Coords, 
        meta: &Self::Meta
    ) -> Option<Vec<usize>>;
}

Trait through which a consumer of this crate sets up the potential to be used.

It is used to initialize the atom types and pair potentials when Lammps is initialized.

The Meta associated type will be the metadata type accepted by e.g. set_structure. Feel free to pick something convenient for your application.

Associated Types

type Meta: Clone

Loading content...
Required methods

fn init_info(&self, coords: &Coords, meta: &Self::Meta) -> InitInfo

Produce information needed by rsp2_lammps_wrap to initialize the potential.

See InitInfo for more information.

Currently, this crate does not support modification of the defined atom types or the potential on an existing Lammps. This method will be called once on the structure used to build a Lammps (to generate the initialization commands), as well as before each computation (to verify that the necessary commands have not changed).

fn atom_types(&self, coords: &Coords, meta: &Self::Meta) -> Vec<AtomType>

Assign atom types to each atom.

The reason this responsibility is deferred to code outside this crate is because different potentials may need to use atom types in wildly different ways. For many potentials, the atom types will simply map to elemental species; but for a potential like ‘kolmogorov/crespi/z’, atom types must be carefully assigned to prevent interactions between non-adjacent layers.

fn molecule_ids(&self, coords: &Coords, meta: &Self::Meta) -> Option<Vec<usize>>

Assign molecule ids to each atom.

This should always return None, or always return Some(_). If it returns None, molecule ids will not be included in the atom style.

Loading content...